+++ src/sasctl/_services/model_repository.py
@@ -9,7 +9,7 @@
 import six
 
 from .service import Service
-from ..core import current_session, get, delete
+from ..core import current_session, get, delete, sasctl_command, HTTPError
 
 FUNCTIONS = {'Analytical', 'Classification', 'Clustering', 'Forecasting',
              'Prediction', 'Text categorization', 'Text extraction',
@@ -38,7 +38,7 @@ class ModelRepository(Service):
 
     _SERVICE_ROOT = '/modelRepository'
 
-    list_repositories, get_repository, update_repository, \
+    list_repositories, _, update_repository, \
         delete_repository = Service._crud_funcs('/repositories', 'repository',
                                                 get_filter=_get_filter)
 
@@ -147,6 +147,67 @@ class ModelRepository(Service):
         link = cls.get_model_link(model, 'contents', refresh=True)
 
         return cls.request_link(link, 'contents')
+
+    @classmethod
+    @sasctl_command('get', 'repositories')
+    def get_repository(cls, repository, refresh=False):
+        """Return a repository instance.
+
+        Parameters
+        ----------
+        repository : str or dict
+            Name, ID, or dictionary representation of the repository.
+        refresh : bool, optional
+            Obtain an updated copy of the repository.
+
+        Returns
+        -------
+        RestObj or None
+            A dictionary containing the repository attributes or None.
+
+        Notes
+        -------
+        If `repository` is a complete representation of the repository it will be
+        returned unless `refresh` is set.  This prevents unnecessary REST
+        calls when data is already available on the client.
+
+        """
+        # If the input already appears to be the requested object just
+        # return it, unless a refresh of the data was explicitly requested.
+        if isinstance(repository, dict) and all(k in repository for k in ('id', 'name')):
+            if refresh:
+                repository = repository['id']
+            else:
+                return repository
+
+        if cls.is_uuid(repository):
+            try:
+                # Attempt to GET the repository directly.  Access may be restricted, so allow HTTP 403 errors
+                # and fall back to using list_repositories() instead.
+                return cls.get('/repositories/{id}'.format(id=repository))
+            except HTTPError as e:
+                if e.code != 403:
+                    raise e
+
+        results = cls.list_repositories()
+
+        # Not sure why, but as of 19w04 the filter doesn't seem to work
+        for result in results:
+            if result['name'] == str(repository) or result['id'] == str(repository):
+                # Make a request for the specific object so that ETag
+                # is included, allowing updates.
+                try:
+                    if cls.get_link(result, 'self'):
+                        return cls.request_link(result, 'self')
+
+                    id_ = result.get('id', result['name'])
+                    return cls.get('/repositories/{id}'.format(id=id_))
+                except HTTPError as e:
+                    # NOTE: As of Viya 4.0.1 access to GET a repository is restricted to admin users out of the box.
+                    # Try to GET the repository, but ignore any 403 (permission denied) errors.
+                    if e.code != 403:
+                        raise e
+                return result
 
     @classmethod
     def create_model(cls, model, project,
@@ -332,15 +393,20 @@ class ModelRepository(Service):
         RestObj
 
         """
-        repo = cls.get_repository('Repository 1')  # Default in 19w04
-        if repo is None:
-            repo = cls.get_repository('Public')  # Default in 19w21
-        if repo is None:
-            all_repos = cls.list_repositories()
-            if all_repos:
-                repo = all_repos[0]
+        all_repos = cls.list_repositories()
 
-        return repo
+        if all_repos:
+            # If nothing else, return the first repository
+            repo = all_repos[0]
+
+            # Check repository names to find a better default.
+            # 'Repository 1' was default in 19w04
+            # 'Public' was default in 19w21
+            for r in all_repos:
+                if r.name in ('Repository 1', 'Public'):
+                    repo = r
+                    break
+            return repo
 
     @classmethod
     def create_project(cls, project, repository, **kwargs):
@@ -386,10 +452,10 @@ class ModelRepository(Service):
         project : str or dict
             The name or id of the model project, or a dictionary
             representation of the project.
-        description : str
-            The description of the model.
         file : bytes
             The ZIP file containing the model and contents.
+        description : str
+            The description of the model.
 
         Returns
         -------